apiserver: update copy, rename v1beta3 to v1
