+++ pkg/apis/flowcontrol/v1/doc.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2022 The Kubernetes Authors.
+Copyright 2023 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -15,10 +15,10 @@ limitations under the License.
 */
 
 // +k8s:conversion-gen=k8s.io/kubernetes/pkg/apis/flowcontrol
-// +k8s:conversion-gen-external-types=k8s.io/api/flowcontrol/v1beta3
+// +k8s:conversion-gen-external-types=k8s.io/api/flowcontrol/v1
 // +k8s:defaulter-gen=TypeMeta
-// +k8s:defaulter-gen-input=k8s.io/api/flowcontrol/v1beta3
+// +k8s:defaulter-gen-input=k8s.io/api/flowcontrol/v1
 
 // +groupName=flowcontrol.apiserver.k8s.io
 
-package v1beta3 // import "k8s.io/kubernetes/pkg/apis/flowcontrol/v1beta3"
+package v1 // import "k8s.io/kubernetes/pkg/apis/flowcontrol/v1"
